Conflicts:
	.classpath
	.gitignore
	.project
	.settings/org.eclipse.jdt.core.prefs
	build.properties
	build.xml
	pom.xml
	src/gov/nasa/jpl/ae/event/DurativeEvent.java
	src/gov/nasa/jpl/ae/event/HasEvents.java
	src/gov/nasa/jpl/ae/event/HasParameters.java
	src/gov/nasa/jpl/ae/event/HasTimeVaryingObjects.java
	src/gov/nasa/jpl/ae/event/TimeVarying.java
	src/gov/nasa/jpl/ae/event/TimeVaryingMap.java
	src/gov/nasa/jpl/ae/magicdrawPlugin/MdDebug.java
	src/gov/nasa/jpl/ae/magicdrawPlugin/modelQuery/JavaModelQueryParser.java
	src/gov/nasa/jpl/ae/solver/CollectionTree.java
	src/gov/nasa/jpl/ae/solver/Constraint.java
	src/gov/nasa/jpl/ae/solver/Domain.java
	src/gov/nasa/jpl/ae/solver/HasConstraints.java
	src/gov/nasa/jpl/ae/solver/HasIdImpl.java
	src/gov/nasa/jpl/ae/solver/Satisfiable.java
	src/gov/nasa/jpl/ae/solver/Variable.java
	src/gov/nasa/jpl/ae/util/ClassData.java

+++ src/gov/nasa/jpl/ae/solver/CollectionTree.java
@@ -687,12 +687,12 @@ import java.util.TreeMap;
       HashSet<Object> h = new HashSet<Object>();
       ct.add(h);
       h.add( "e" );
-      h.add( 2.71 );
+      h.add( (Double)2.71 );
       h.add( (double)9.8 );
       TreeMap<String,Object> tm = new TreeMap< String, Object >();
       h.add( tm );
       tm.put( "foo", "bar" );
-      tm.put( "gravity", new Float(9.8) );
+      tm.put( "gravity", new Double(9.8) );
       java.util.Iterator i = ct.iterator();
       System.out.println("ct = " + ct );
     }
